fix: Move to a11y-folder, shared onKeyDown test method
