+++ src/a11y/index.js
@@ -1,0 +1,1 @@
+export { getAriaLabel } from './a11y.utils'
