+++ src/index.js
@@ -10,7 +10,8 @@ import cn from 'classnames/bind'
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 
-import { isOutsideClick, clientIdGenerator, getAriaLabel } from './utils'
+import { isOutsideClick, clientIdGenerator } from './utils'
+import { getAriaLabel } from './a11y'
 import Input from './input'
 import Tree from './tree'
 import TreeManager from './tree-manager'
