+++ src/utils/index.js
@@ -4,4 +4,3 @@ export { default as isEmpty } from './isEmpty'
 export { default as isOutsideClick } from './isOutsideClick'
 export { default as debounce } from './debounce'
 export { default as clientIdGenerator } from './generateUID'
-export { getAriaLabel } from './a11yHelper'
